elearn/requirements.txt
Django
Pillow
requests
django-debug-toolbar
django-jazzmin
channels 
channels_redis 
djangorestframework
